Fix servo examples.

+++ stm32f0_hal/examples/servo.rs
@@ -13,12 +13,12 @@ fn main() {
     loop {
         // increase position of 0.5° each 20 ms
         for i in 0..360 {
-            servo1.set_position(i as f32 / 2.0);
+            servo1.set_position(i as f64 / 2.0);
             rcc::ms_delay(20);
         }
         // decrease position of 0.5° each 20 ms
         for i in 0..360 {
-            servo1.set_position(180.0 - (i as f32 / 2.0));
+            servo1.set_position(180.0 - (i as f64 / 2.0));
             rcc::ms_delay(20);
         }
     }
